chore: better error handling, test and code review
