+++ crates/rome_cli/src/commands/ci.rs
@@ -9,8 +9,7 @@ use super::format::parse_format_options;
 
 /// Handler for the "ci" command of the Rome CLI
 pub(crate) fn ci(mut session: CliSession) -> Result<(), Termination> {
-    // reading the configuration should not cause an error, rome should working even without it
-    let configuration = load_config(&session.app.fs, ConfigurationType::Root).unwrap_or(None);
+    let configuration = load_config(&session.app.fs, ConfigurationType::Root)?;
 
     parse_format_options(&mut session, configuration)?;
     traverse(TraversalMode::CI, session)
