+++ crates/rome_cli/tests/main.rs
@@ -1,3 +1,5 @@
+mod configs;
+
 use std::{ffi::OsString, path::Path};
 
 use pico_args::Arguments;
@@ -38,25 +40,6 @@ return { something }
 // six spaces
 const CUSTOM_FORMAT_AFTER: &str = r#"function f() {
       return { something };
-}
-"#;
-
-const CONFIG_FORMAT: &str = r#"{
-  "root": true,
-  "formatter": {
-    "enabled": true,
-    "lineWidth": 160,
-    "indentStyle": "space",
-    "indentSize": 6
-  }
-}
-"#;
-
-const CONFIG_DISABLED_FORMATTER: &str = r#"{
-  "root": true,
-  "formatter": {
-    "enabled": false
-  }
 }
 "#;
 
@@ -287,7 +270,7 @@ mod ci {
 
 mod format {
     use super::*;
-    use std::env::current_dir;
+    use crate::configs::{CONFIG_DISABLED_FORMATTER, CONFIG_FORMAT};
 
     #[test]
     fn print() {
@@ -507,8 +490,7 @@ mod format {
     #[test]
     fn format_with_configuration() {
         let mut fs = MemoryFileSystem::default();
-        let config_path = current_dir().unwrap().join("rome.json");
-        let file_path = Path::new(config_path.as_os_str());
+        let file_path = Path::new("rome.json");
         fs.insert(file_path.into(), CONFIG_FORMAT.as_bytes());
 
         let file_path = Path::new("file.js");
@@ -542,8 +524,7 @@ mod format {
     #[test]
     fn format_is_disabled() {
         let mut fs = MemoryFileSystem::default();
-        let config_path = current_dir().unwrap().join("rome.json");
-        let file_path = Path::new(config_path.as_os_str());
+        let file_path = Path::new("rome.json");
         fs.insert(file_path.into(), CONFIG_DISABLED_FORMATTER.as_bytes());
 
         let file_path = Path::new("file.js");
@@ -667,6 +648,87 @@ mod main {
         match result {
             Err(Termination::MissingArgument { argument }) => assert_eq!(argument, "<INPUT>"),
             _ => panic!("run_cli returned {result:?} for a missing argument, expected an error"),
+        }
+    }
+}
+
+mod configuration {
+    use crate::configs::{CONFIG_ALL_FIELDS, CONFIG_BAD_LINE_WIDTH, CONFIG_ROOT_FALSE};
+    use pico_args::Arguments;
+    use rome_cli::{run_cli, CliSession};
+    use rome_console::BufferConsole;
+    use rome_fs::MemoryFileSystem;
+    use rome_service::{App, DynRef};
+    use std::ffi::OsString;
+    use std::path::Path;
+
+    #[test]
+    fn incorrect_root() {
+        let mut fs = MemoryFileSystem::default();
+        let file_path = Path::new("rome.json");
+        fs.insert(file_path.into(), CONFIG_ROOT_FALSE.as_bytes());
+
+        let result = run_cli(CliSession {
+            app: App::with_filesystem_and_console(
+                DynRef::Borrowed(&mut fs),
+                DynRef::Owned(Box::new(BufferConsole::default())),
+            ),
+            args: Arguments::from_vec(vec![OsString::from("format"), OsString::from("file.js")]),
+        });
+
+        assert!(result.is_err());
+
+        match result {
+            Err(error) => {
+                assert_eq!(
+                    error.to_string(),
+                    "the main configuration file, rome.json, must have the field 'root' set to `true`"
+                )
+            }
+            _ => panic!("expected an error, but found none"),
+        }
+    }
+
+    #[test]
+    fn correct_root() {
+        let mut fs = MemoryFileSystem::default();
+        let file_path = Path::new("rome.json");
+        fs.insert(file_path.into(), CONFIG_ALL_FIELDS.as_bytes());
+
+        let result = run_cli(CliSession {
+            app: App::with_filesystem_and_console(
+                DynRef::Borrowed(&mut fs),
+                DynRef::Owned(Box::new(BufferConsole::default())),
+            ),
+            args: Arguments::from_vec(vec![OsString::from("format"), OsString::from("file.js")]),
+        });
+
+        assert!(result.is_ok());
+    }
+
+    #[test]
+    fn line_width_error() {
+        let mut fs = MemoryFileSystem::default();
+
+        let file_path = Path::new("rome.json");
+        fs.insert(file_path.into(), CONFIG_BAD_LINE_WIDTH.as_bytes());
+
+        let result = run_cli(CliSession {
+            app: App::with_filesystem_and_console(
+                DynRef::Borrowed(&mut fs),
+                DynRef::Owned(Box::new(BufferConsole::default())),
+            ),
+            args: Arguments::from_vec(vec![OsString::from("format"), OsString::from("file.js")]),
+        });
+        assert!(result.is_err());
+
+        match result {
+            Err(error) => {
+                assert!(error
+                    .to_string()
+                    .contains("The line width exceeds the maximum value (320)"),)
+            }
+            _ => panic!("expected an error, but found none"),
         }
     }
 }
