--- crates/rome_service/tests/all_fields.json
@@ -1,15 +1,0 @@
-{
-  "root": true,
-  "formatter": {
-    "enabled": true,
-    "formatWithErrors": true,
-    "indentStyle": "tab",
-    "indentSize": 2,
-    "lineWidth": 80
-  },
-  "javascript": {
-    "formatter": {
-      "quoteStyle": "double"
-    }
-  }
-}
