--- crates/rome_service/tests/empty.json
@@ -1,1 +1,0 @@
-{}
