--- crates/rome_service/tests/line_width.json
@@ -1,6 +1,0 @@
-{
-  "root": true,
-  "formatter": {
-    "lineWidth": 450
-  }
-}
